+++ tools/src/main/java/org/esa/cci/sst/regavg/ProductType.java
@@ -3,10 +3,11 @@ package org.esa.cci.sst.regavg;
 import org.esa.cci.sst.util.GridDef;
 import org.esa.cci.sst.util.UTC;
 
+import java.io.File;
+import java.io.IOException;
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.util.Date;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 /**
@@ -23,18 +24,23 @@ public enum ProductType {
      */
     ARC_L3U {
         public final DateFormat dateFormat = UTC.getDateFormat("yyyyMMdd");
-        public final int filenameDateOffset = "AT._AVG_3PAARC".length();
-        public final Pattern filenamePattern = Pattern.compile("AT._AVG_3PAARC\\d\\d\\d\\d\\d\\d\\d.*\\.nc.gz");
+        public final int filenameDateOffset = "ATS_AVG_3PAARC".length();
+        public final String filenameRegex = "AT[12S]_AVG_3PAARC\\d{8}_[DTEM]_[nd][ND][23][bms][.]nc[.]gz";
         public final GridDef GRID_DEF = GridDef.createGlobalGrid(3600, 1800);
 
         @Override
-        public Date getDate(String filename) throws ParseException {
-            Matcher matcher = filenamePattern.matcher(filename);
-            if (matcher.matches()) {
-                String dateString = filename.substring(filenameDateOffset, filenameDateOffset + 8);
+        public Date getDate(File file) throws IOException {
+            try {
+                String dateString = file.getName().substring(filenameDateOffset, filenameDateOffset + 8);
                 return dateFormat.parse(dateString);
+            } catch (ParseException e) {
+                throw new IOException("Illegal '" + ARC_L3U + "' filename: " + file);
             }
-            return null;
+        }
+
+        @Override
+        public String getDefaultFilenameRegex() {
+            return filenameRegex;
         }
 
         @Override
@@ -50,13 +56,18 @@ public enum ProductType {
 
     ARC_L2P {
         @Override
-        public Date getDate(String filename) {
-            return null;
+        public Date getDate(File file) {
+            throw new IllegalStateException("Not implemented.");
         }
 
         @Override
         public GridDef getGridDef() {
-            return null;
+            throw new IllegalStateException("Not implemented.");
+        }
+
+        @Override
+        public String getDefaultFilenameRegex() {
+            throw new IllegalStateException("Not implemented.");
         }
 
         @Override
@@ -67,14 +78,19 @@ public enum ProductType {
 
     CCI_L3U {
         @Override
-        public Date getDate(String filename) {
-            return null;
+        public Date getDate(File file) {
+            throw new IllegalStateException("Not implemented.");
         }
 
 
         @Override
         public GridDef getGridDef() {
-            return null;
+            throw new IllegalStateException("Not implemented.");
+        }
+
+        @Override
+        public String getDefaultFilenameRegex() {
+            throw new IllegalStateException("Not implemented.");
         }
 
         @Override
@@ -85,14 +101,19 @@ public enum ProductType {
 
     CCI_L3C {
         @Override
-        public Date getDate(String filename) {
-            return null;
+        public Date getDate(File file) {
+            throw new IllegalStateException("Not implemented.");
         }
 
 
         @Override
         public GridDef getGridDef() {
-            return null;
+            throw new IllegalStateException("Not implemented.");
+        }
+
+        @Override
+        public String getDefaultFilenameRegex() {
+            throw new IllegalStateException("Not implemented.");
         }
 
         @Override
@@ -103,14 +124,19 @@ public enum ProductType {
 
     CCI_L4 {
         @Override
-        public Date getDate(String filename) {
-            return null;
+        public Date getDate(File file) {
+            throw new IllegalStateException("Not implemented.");
         }
 
 
         @Override
         public GridDef getGridDef() {
-            return null;
+            throw new IllegalStateException("Not implemented.");
+        }
+
+        @Override
+        public String getDefaultFilenameRegex() {
+            throw new IllegalStateException("Not implemented.");
         }
 
         @Override
@@ -119,7 +145,9 @@ public enum ProductType {
         }
     };
 
-    public abstract Date getDate(String filename) throws ParseException;
+    public abstract Date getDate(File file) throws IOException;
+
+    public abstract String getDefaultFilenameRegex();
 
     public abstract GridDef getGridDef();